add swagger

+++ src/pessoas/pessoas.service.ts
@@ -1,5 +1,7 @@
 import {
+  BadRequestException,
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -7,13 +9,20 @@ import { CreatePessoaDto } from './dto/create-pessoa.dto';
 import { UpdatePessoaDto } from './dto/update-pessoa.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Pessoa } from './entities/pessoa.entity';
-import { Repository } from 'typeorm';
+import { Not, Repository } from 'typeorm';
+import { RecadosUtils } from 'src/recados/recados.utils';
+import { HashingService } from 'src/auth/hashing/hashing.service';
+import { TokenPayloadDto } from 'src/auth/dto/token-payload.dto';
+import * as path from 'path';
+import * as fs from 'fs/promises';
 
 @Injectable()
 export class PessoasService {
   constructor(
     @InjectRepository(Pessoa)
     private readonly pessoaRepository: Repository<Pessoa>,
+    private readonly recadosUtils: RecadosUtils,
+    private readonly hashingService: HashingService,
   ) {}
 
   notFoundPerson<T>(
@@ -27,10 +36,15 @@ export class PessoasService {
 
   async create(createPessoaDto: CreatePessoaDto) {
     try {
+      const passwordHash = await this.hashingService.hash(
+        createPessoaDto.password,
+      );
+
       const pessoaData = {
         nome: createPessoaDto.nome,
-        passwordHash: createPessoaDto.password,
+        passwordHash,
         email: createPessoaDto.email,
+        routePolicies: createPessoaDto.routePolicies,
       };
 
       const novaPessoa = this.pessoaRepository.create(pessoaData);
@@ -58,18 +72,33 @@ export class PessoasService {
   }
 
   async findOne(id: number) {
-    const pessoa = await this.pessoaRepository.findOneBy({
-      id,
-    });
+    // console.log(this.recadosUtils.inverteString('Luiz'));
+
+    const pessoa = await this.pessoaRepository.findOneBy({ id });
+
+    if (!pessoa) {
+      throw new NotFoundException('Pessoa não encontrada');
+    }
 
     return pessoa;
   }
 
-  async update(id: number, updatePessoaDto: UpdatePessoaDto) {
+  async update(
+    id: number,
+    updatePessoaDto: UpdatePessoaDto,
+    tokenPayload: TokenPayloadDto,
+  ) {
     const pessoaData = {
       nome: updatePessoaDto.nome,
-      passwordHash: updatePessoaDto.password,
     };
+
+    if (updatePessoaDto?.password) {
+      const passwordHash = await this.hashingService.hash(
+        updatePessoaDto.password,
+      );
+
+      pessoaData['passwordHash'] = passwordHash;
+    }
 
     const pessoa = await this.pessoaRepository.preload({
       id,
@@ -78,20 +107,70 @@ export class PessoasService {
 
     this.notFoundPerson(pessoa, 'Pessoa não encontrada');
 
+    if (pessoa?.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Você não é essa pessoa');
+    }
+
     return this.pessoaRepository.save(pessoa);
   }
 
-  async remove(id: number) {
-    const pessoa = await this.pessoaRepository.findOneBy({
-      id,
-    });
+  async remove(id: number, tokenPayload: TokenPayloadDto) {
+    const pessoa = await this.findOne(id);
+
+    if (pessoa?.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Você não é essa pessoa');
+    }
+
+    return this.pessoaRepository.remove(pessoa);
+  }
+
+  async uploadPicture(
+    file: Express.Multer.File,
+    tokenPayload: TokenPayloadDto,
+  ) {
+    if (file.size < 1024) {
+      throw new BadRequestException('File too small');
+    }
+
+    // 🔴 PARA VÁRIOS ARQUIVOS:
+    // const result: string[] = [];
+    // files.forEach(async file => {
+    //   const fileExtension = path
+    //     .extname(file.originalname)
+    //     .toLowerCase()
+    //     .substring(1);
+    //   const fileName = `${randomUUID()}.${fileExtension}`;
+    //   const fileFullPath = path.resolve(process.cwd(), 'pictures', fileName);
+    //   console.log(fileFullPath);
+
+    //   result.push(fileFullPath);
+
+    //   await fs.writeFile(fileFullPath, file.buffer);
+    // });
+    // return result;
+
+    // 🔴 PARA UM ÚNICO ARQUIVO:
+    const pessoa = await this.findOne(tokenPayload.sub);
 
     // if (!pessoa) {
     //   throw new NotFoundException('Pessoa não encontrada');
     // }
 
-    this.notFoundPerson(pessoa, 'Pessoa não encontrada');
+    const fileExtension = path
+      .extname(file.originalname)
+      .toLowerCase()
+      .substring(1);
+    const fileName = `${tokenPayload.sub}.${fileExtension}`;
+    const fileFullPath = path.resolve(process.cwd(), 'pictures', fileName);
+    console.log(fileFullPath);
 
-    return this.pessoaRepository.remove(pessoa);
+    // 🧑‍🎓❗ pesquisar e estudar sobre essas bibliotecas: file-type image-type sharp
+
+    await fs.writeFile(fileFullPath, file.buffer);
+
+    pessoa.picture = fileName;
+    await this.pessoaRepository.save(pessoa);
+
+    return pessoa;
   }
 }
